initial implementation of SigChain
This implementation includes a test sigchain that I generated by hand
with a test user, which the code is able to parse and validate. Nothing
is there yet fetching from the Merkle tree; the code simply accepts the
UID, username, and eldest key as inputs.

Once we get this properly tested using Max's node-forge-sigchain, and
the API really nailed down, I'll make everything pretty and add some
more comments.