stop calling node-forge-sigchain directly
The goal will be to check in output from that tool and just test on that
output. Eventually we might use another repo, but for now I'm just
checking it in here.

Note that we have to hardcode the username, uid, and eldest key. It
would be good to change that somehow.