update the sigchain interface
ParsedKeys now markes the key map as private, and only exposes a get
method. I noticed myself making the mistake of using the collection of
parsed keys as a standin for the user's valid sibkeys, which is a major
mistake, and hopefully this more restrictive interface will prevent it
in the future.

get_sibkeys() and get_subkeys() now return a list of key managers,
rather than just KIDs. This should save callers from needing to keep the
ParsedKeys object around. The keys are given in the order they're
delegated.