stop using server ctime
We've discovered that the server sometimes has bad values for the ctimes
of some old links. Rather than guarantee that the server always has this
right, we've decided to stop using the server ctime entirely, and just
get the signing ctime straight from the unverified sig payload.

TODO: Pin the new version of kbpgp which supports this.