allow repeating eldest keys in the sigchain, and respect 'eldest' link resets
Before, we used to short-circuit a lot of our sigchain replay code until
we got to the current subchain. Now, we actually do a complete replay of
all subchains, and implement a _reset_subchain() method to clear the
relevant state when we detect a reset.

We need this logic because now there might be more than one subchain
that corresponds to our current eldest key. The only other option
would've been to detect the current subchain on a first pass, and only
then do a full replay. But that makes it difficult to keep track of the
KeyState that tells us what PGP full hash to use at any given time.
Fully replaying everything is actually simpler, and our signature cache
hopefully absolves us of any major perf worries.